@@ -1,10 +1,10 @@
 {
-  "name": "@theia/node-pty",
+  "name": "node-pty",
   "description": "Fork pseudoterminals in Node.JS",
   "author": {
     "name": "Microsoft Corporation"
   },
-  "version": "0.8.1-theia.5",
+  "version": "0.9.0-theia.6",
   "license": "MIT",
   "main": "./lib/index.js",
   "types": "./typings/node-pty.d.ts",
@@ -39,21 +39,20 @@
     "install": "node scripts/install.js",
     "postinstall": "node scripts/post-install.js",
     "test": "cross-env NODE_ENV=test mocha -R spec --exit lib/*.test.js",
-    "prepublish": "npm run tsc"
+    "prepare": "npm run tsc",
+    "prepublishOnly": "npm run tsc"
   },
   "dependencies": {
-    "nan": "2.12.1"
+    "nan": "^2.14.0"
   },
   "devDependencies": {
     "@types/mocha": "^5.0.0",
-    "@types/node": "^6.0.104",
-    "@types/ps-list": "^6.0.0",
+    "@types/node": "8",
     "cross-env": "^5.1.4",
     "mocha": "^5.0.5",
-    "pollUntil": "^1.0.3",
     "ps-list": "^6.0.0",
     "tslint": "^5.12.1",
     "tslint-consistent-codestyle": "^1.15.0",
-    "typescript": "^3.2.2"
+    "typescript": "3.4"
   }
 }
