@@ -6,9 +6,13 @@
  * single console attached to a process.
  */
 
-import { loadNative } from './utils';
+let getConsoleProcessList: any;
+try {
+  getConsoleProcessList = require('../build/Release/conpty_console_list.node').getConsoleProcessList;
+} catch (err) {
+  getConsoleProcessList = require('../build/Debug/conpty_console_list.node').getConsoleProcessList;
+}
 
-const getConsoleProcessList = loadNative('conpty_console_list').getConsoleProcessList;
 const shellPid = parseInt(process.argv[2], 10);
 const consoleProcessList = getConsoleProcessList(shellPid);
 process.send({ consoleProcessList });
