@@ -13,7 +13,7 @@
 
 namespace path_util {
 
-const wchar_t* to_wstring(const v8::String::Utf8Value& str);
+const wchar_t* to_wstring(const Nan::Utf8String& str);
 bool file_exists(std::wstring filename);
 std::wstring get_shell_path(std::wstring filename);
 
