@@ -7,7 +7,7 @@
 import { Socket } from 'net';
 import { Terminal, DEFAULT_COLS, DEFAULT_ROWS } from './terminal';
 import { WindowsPtyAgent } from './windowsPtyAgent';
-import { IPtyForkOptions, IPtyOpenOptions } from './interfaces';
+import { IPtyOpenOptions, IWindowsPtyForkOptions } from './interfaces';
 import { ArgvOrCommandLine } from './types';
 import { assign } from './utils';
 
@@ -19,7 +19,7 @@ export class WindowsTerminal extends Terminal {
   private _deferreds: any[];
   private _agent: WindowsPtyAgent;
 
-  constructor(file?: string, args?: ArgvOrCommandLine, opt?: IPtyForkOptions) {
+  constructor(file?: string, args?: ArgvOrCommandLine, opt?: IWindowsPtyForkOptions) {
     super(opt);
 
     // Initialize arguments
@@ -33,8 +33,8 @@ export class WindowsTerminal extends Terminal {
     }
 
     const env = assign({}, opt.env);
-    const cols = opt.cols || DEFAULT_COLS;
-    const rows = opt.rows || DEFAULT_ROWS;
+    this._cols = opt.cols || DEFAULT_COLS;
+    this._rows = opt.rows || DEFAULT_ROWS;
     const cwd = opt.cwd || process.cwd();
     const name = opt.name || env.TERM || DEFAULT_NAME;
     const parsedEnv = this._parseEnv(env);
@@ -46,7 +46,7 @@ export class WindowsTerminal extends Terminal {
     this._deferreds = [];
 
     // Create new termal.
-    this._agent = new WindowsPtyAgent(file, args, parsedEnv, cwd, cols, rows, false, opt.experimentalUseConpty);
+    this._agent = new WindowsPtyAgent(file, args, parsedEnv, cwd, this._cols, this._rows, false, opt.useConpty, opt.conptyInheritCursor);
     this._socket = this._agent.outSocket;
 
     // Not available until `ready` event emitted.
@@ -118,9 +118,19 @@ export class WindowsTerminal extends Terminal {
     this._readable = true;
     this._writable = true;
 
+    this._forwardEvents();
+
     process.nextTick(() => {
       this.emit('exec', undefined);
     });
+  }
+
+  protected _write(data: string): void {
+    this._defer(this._doWrite, data);
+  }
+
+  private _doWrite(data: string): void {
+    this._agent.inSocket.write(data);
   }
 
   /**
@@ -132,25 +142,17 @@ export class WindowsTerminal extends Terminal {
   }
 
   /**
-   * Events
-   */
-
-  public write(data: string): void {
-    this._defer(() => {
-      this._agent.inSocket.write(data);
-    });
-  }
-
-  /**
    * TTY
    */
 
   public resize(cols: number, rows: number): void {
-    if (cols <= 0 || rows <= 0) {
+    if (cols <= 0 || rows <= 0 || isNaN(cols) || isNaN(rows) || cols === Infinity || rows === Infinity) {
       throw new Error('resizing must be done using positive cols and rows');
     }
     this._defer(() => {
       this._agent.resize(cols, rows);
+      this._cols = cols;
+      this._rows = rows;
     });
   }
 
@@ -170,22 +172,16 @@ export class WindowsTerminal extends Terminal {
     });
   }
 
-  private _defer(deferredFn: Function): void {
-
-    // Ensure that this method is only used within Terminal class.
-    if (!(this instanceof WindowsTerminal)) {
-      throw new Error('Must be instanceof WindowsTerminal');
-    }
-
+  private _defer<A extends any>(deferredFn: (arg?: A) => void, arg?: A): void {
     // If the terminal is ready, execute.
     if (this._isReady) {
-      deferredFn.apply(this, null);
+      deferredFn.call(this, arg);
       return;
     }
 
     // Queue until terminal is ready.
     this._deferreds.push({
-      run: () => deferredFn.apply(this, null)
+      run: () => deferredFn.call(this, arg)
     });
   }
 
